refactor: Moved initCommand to init.js
- Moved was needed because if initCommand was called during tests,
index.js was loaded, breaking the execution
